project/requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==1.11.15
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.8.2
docopt==0.6.2
idna==2.7
mongoengine==0.15.3
pipreqs==0.4.9
pymongo==3.7.1
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
yarg==0.1.9
